Move BBox.ts to common directory
